fix syntax error of extra period in R versioning
Closes #8 

https://github.com/binder-examples/r#requirements-and-suggestions

Requirements and suggestions
You need to have a runtime.txt file that is formatted like:

r-<YYYY>-<MM>-<DD>
where YYYY-MM-DD is a snapshot at MRAN that will be used for installing libraries. In this line, you can request a specific version of R. To do this list the version between the 'r' and the year, as in r-3.6-2019-09-24. Right now the default version of R is 3.6.